// apps/server/src/modules/common-cartridge/common-cartridge-client/room-client/room-api-client/models/board-element-response.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Schulcloud-Verbund-Software Server API
 * This is v3 of Schulcloud-Verbund-Software Server. Checkout /docs for v1.
 *
 * The version of the OpenAPI document: 3.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import type { BoardElementResponseContent } from './board-element-response-content';

/**
 * 
 * @export
 * @interface BoardElementResponse
 */
export interface BoardElementResponse {
    /**
     * the type of the element in the content. For possible types, please refer to the enum
     * @type {string}
     * @memberof BoardElementResponse
     */
    'type': BoardElementResponseType;
    /**
     * 
     * @type {BoardElementResponseContent}
     * @memberof BoardElementResponse
     */
    'content': BoardElementResponseContent;
}

export const BoardElementResponseType = {
    TASK: 'task',
    LESSON: 'lesson',
    COLUMN_BOARD: 'column-board'
} as const;

export type BoardElementResponseType = typeof BoardElementResponseType[keyof typeof BoardElementResponseType];