// apps/server/src/modules/common-cartridge/common-cartridge-client/room-client/room-api-client/models/single-column-board-response.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Schulcloud-Verbund-Software Server API
 * This is v3 of Schulcloud-Verbund-Software Server. Checkout /docs for v1.
 *
 * The version of the OpenAPI document: 3.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import type { BoardElementResponse } from './board-element-response';

/**
 * 
 * @export
 * @interface SingleColumnBoardResponse
 */
export interface SingleColumnBoardResponse {
    /**
     * The id of the room this board belongs to
     * @type {string}
     * @memberof SingleColumnBoardResponse
     */
    'roomId': string;
    /**
     * Title of the Board
     * @type {string}
     * @memberof SingleColumnBoardResponse
     */
    'title': string;
    /**
     * Color of the Board
     * @type {string}
     * @memberof SingleColumnBoardResponse
     */
    'displayColor': string;
    /**
     * Array of board specific tasks or lessons with matching type property
     * @type {Array<BoardElementResponse>}
     * @memberof SingleColumnBoardResponse
     */
    'elements': Array<BoardElementResponse>;
    /**
     * Boolean if the room this board belongs to is archived
     * @type {boolean}
     * @memberof SingleColumnBoardResponse
     */
    'isArchived': boolean;
    /**
     * Is the course synchronized with a group?
     * @type {boolean}
     * @memberof SingleColumnBoardResponse
     */
    'isSynchronized': boolean;
}